Refactor raw mode input to use termios directly
This drops curses for termios, fnctl, and os functions. They are
exactly as unportable as curses itself, because this code does
what curses does to provide raw input, without also messing with
the output.

This allows tqdm, logging, print, scrollback etc to function
normally. A logging handler is implemented to ensure that log
messages interact correctly with tqdm.