Implement function macros
A function macro is a function which yields states. Unlike a file,
it can run forever. This can be used to make autofire/mash macros
which play forever. Example:

a mash 20 00000200000000

This will hold the A button for 21 interrupts then release it
for 21 interrupts, then repeat forever.. All other inputs will
be from the live controller due to the mask.

This commit also does some housekeeping with logging and implements
hexstr on State for easier printing.