Implement SDL window and move global recording out of MacroManager
The SDL window can be used to get raw keyboard input. If windowing
is not available the window will not open and keyboard input will
not be available.

Global recording (-R) is removed from MacroManager into a separate
context manager. This will simplify refactoring MacroManager for
different macro triggering methods.