Re-implement macro triggering and blending
Macros are now triggered by pressing space. If a macro is recording
it will stop recording. Otherwise, the current button and hat state
is checked. The macro is slected based on this state. Then:

1. If the macro exists it starts playing. If a macro was already
playing it will be stopped.

2. If the macro does not exist it will start recording. If a macro
was already playing it will continue playing and be recorded in the
new macro.

When a macro is played it will be blended as follows:

1. User buttons are are ORed with macro buttons.
2. User hat overrides macro input if the user hat is not centered.
3. User axes override macro axes if they are more than 50%.