Implement direct USB mode for Linux UDC, functionfs, and configfs
This enables direct USB mode on Raspberry Pi and other USB device
capable systems. To use it, pass "-p gadget -u <UDC>" where UDC is
the name of a directory in /sys/class/udc. For Raspberry Pi it
should be something like "20980000.usb" and for local loopback
testing it should be "dummy_udc.0"

There are still some problems with this: the report format is wrong
endian so the firmware will have to be changed again to match.